test: Fixed build error in test project and added extra test method
