+++ .editorconfig
@@ -19,3 +19,10 @@ indent_size = 4
 
 [{Makefile,*.mk,go.mod,go.sum,*.go,.gitmodules}]
 indent_style = tab
+
+[*.cs]
+indent_size = 4
+
+[*.sln]
+indent_size = 1
+indent_style = tab
