+++ src/EpisodeManagementService/UpdateEpisode/UpdateEpisode.cs
@@ -7,40 +7,35 @@ using Microsoft.Extensions.Logging;
 using System.Text;
 using System.Text.Json;
 
-
 public class UpdateEpisode
 {
-  private readonly ILogger<UpdateEpisode> _logger;
+    private readonly ILogger<UpdateEpisode> _logger;
 
-  public UpdateEpisode(ILogger<UpdateEpisode> logger)
-  {
-    _logger = logger;
-  }
-
-  [Function("updateEpisode")]
-  public HttpResponseData Run([HttpTrigger(AuthorizationLevel.Anonymous, "get", "post")] HttpRequestData req)
-  {
-    Episode episode;
-    try
+    public UpdateEpisode(ILogger<UpdateEpisode> logger)
     {
-      using (StreamReader reader = new StreamReader(req.Body, Encoding.UTF8))
-      {
-        var postData = reader.ReadToEnd();
-        episode = JsonSerializer.Deserialize<Episode>(postData);
-      }
-
-      _logger.LogInformation(episode.EpisodeId);
-
-      return req.CreateResponse(HttpStatusCode.OK);
-
+        _logger = logger;
     }
-    catch
+
+    [Function("UpdateEpisode")]
+    public HttpResponseData Run([HttpTrigger(AuthorizationLevel.Anonymous, "get", "post")] HttpRequestData req)
     {
-      _logger.LogError("Could not read episode data.");
+        Episode episode;
 
-      return req.CreateResponse(HttpStatusCode.BadRequest);
+        try
+        {
+            using (StreamReader reader = new StreamReader(req.Body, Encoding.UTF8))
+            {
+                var postData = reader.ReadToEnd();
+                episode = JsonSerializer.Deserialize<Episode>(postData);
+            }
+
+            _logger.LogInformation(episode.EpisodeId);
+            return req.CreateResponse(HttpStatusCode.OK);
+        }
+        catch
+        {
+            _logger.LogError("Could not read episode data.");
+            return req.CreateResponse(HttpStatusCode.BadRequest);
+        }
     }
-  }
 }
-
-
