+++ tests/TestUtils/TestUtils.csproj
@@ -22,10 +22,5 @@
   <ItemGroup>
     <Using Include="Microsoft.VisualStudio.TestTools.UnitTesting" />
   </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\application\CohortManager\src\Functions\Shared\Common\Common.csproj" />
-  </ItemGroup>
-
-
+  
 </Project>
